some fix with poemForm

+++ migrations/m161005_134556_create_table_poem.php
@@ -13,7 +13,7 @@ class m161005_134556_create_table_poem extends Migration
             'poem' => $this->text()->notNull(),
             'autor' => $this->string(100)->notNull(),
             'date' => $this->integer()->notNull(),
-            'censor' => $this->integer(1)->notNull(),
+            'censor' => $this->integer()->notNull(),
         ]);
 
     }
